delete buttons work

+++ javascripts/db-interaction.js
@@ -43,7 +43,7 @@ function getMovies(searchResult){
 function addToMyMovies() {
     console.log('you clicked I want to see this movie');
     var currentCard = $(event.currentTarget);
-    console.log('url:', currentCard.siblings("img").attr("src"));
+    // console.log('url:', currentCard.siblings("img").attr("src"));
     var currentUser = user.getUser();
     var myMovie = {
         "title": currentCard.siblings("h3").html(),
@@ -66,15 +66,28 @@ function addToMyMovies() {
 }
 
 // Deletes a movie using the movie's UID
-function deleteMovie(movieID){
-  return new Promise( function(resolve, reject){
+function deleteMovie(movieTitle){
+    // console.log('movieTitle = ', movieTitle);
+    var currentUser = user.getUser();
     $.ajax({
-      url: `https://movie-history-6e707.firebaseio.com/movies/${movieID}.json`,
-      method: "DELETE"
-    }).done( function(){
-      resolve();
+        url: `https://movie-history-6e707.firebaseio.com/movies.json`,
+        type: "GET"
+    }).done(function(movieData){
+        // console.log('movieData = ', movieData);
+        var moviesKeys = Object.keys(movieData);
+        var moviesObjects = Object.values(movieData);
+        // console.log('movies in deleteMovie = ', movies);
+        for (var i = 0; i < moviesObjects.length; i++) {
+            if (currentUser === moviesObjects[i].userID && movieTitle === moviesObjects[i].title) {
+                console.log('movie that matches userID and title of delete button = ', moviesKeys[i]);
+                $.ajax({
+                    url: `https://movie-history-6e707.firebaseio.com/movies/${moviesKeys[i]}.json`,
+                    method: "DELETE"
+                });
+            }
+        }
+
     });
-  });
 }
 
 function searchFirebase(searchString){
